created a separate command for initializing drive in auto
The AutoRoutineCommand would not work if a command group was passed into it and the command group contained the first FollowCommand. I also realized that there were other command groups than the one AutoRoutineCommand extended (SequentialCommandGroup). Fixing these issues would require a lot of additional code, so in the end I decided it was simpler to just create a command that initialized the drive subsystem for autonomous without looking for the first FollowCommand. Instead, the new command is given the data needed to generate the first followed trajectory, from which it can get the initial pose.