testing, debugging, and setup
We tested the straight test auto routine and it went straight (for the most part), but did whatever it wanted afterwards. To fix this, we overrode the end() method in FollowCommand.java to tell the drive subsystem to stop. This fixed the issue. However, some PID tuning is necessary. The required SmartDashboard entries have been set up in code.