changing to use holonomic rotation on init
Previously, the initial pose of the trajectory was being used to for drive subsystem auto initialization, but it turns out that the poses of the trajectory do not reflect the holonomic rotations of the trajectory (the actual heading of the robot as it follows the trajectory). Now, a new pose is created with the initial pose's translation but the initial holonomic rotation, and that is used for drive subsystem auto intialization.