more PID tuning (not complete)
After running the strafe test, we found that the x controller and y controller do not necessarily have the same parameters. As such, we will retune the x controller and y controller.

After some experimentation, the method I came up with was as follows:
1. Get the highest P value that does not cause anything funny (oscillation, back-tracking, etc) for the x controller using the straight test.
2. Get the highest P value that does not cause anything funny (oscillation, back-tracking, etc) for the y controller using the strafe test.
3. Tweak the x controller and y controller by following a curved path, tweaking the parameters to produce optimal behavior.

Note: we do not know if this method will work as the resulting parameters might not work for something simple like straight test or strafe test in terms of accuracy.

We also fixed crook in advanced_test_2 path caused by a waypoint's heading being off.