Update RotateCommand.java
Issue with RotateCommands where the setpoint would be based on robot's rotation during robot bootup (this was because the setpoint was being calculated in the constructor of the RotateCommands, which were being created during robot initialization). Solved by moving setpoint calculation to the initialize block of the RotateCommand (which runs when the command is scheduled, which is not during robot bootup).